+++ tensorflow/compiler/xla/service/gpu/nvptx_compiler.cc
@@ -55,7 +55,6 @@ limitations under the License.
 #include "tensorflow/core/platform/cuda_libdevice_path.h"
 #include "tensorflow/core/platform/tracing.h"
 #include "tensorflow/core/profiler/lib/traceme.h"
-#include "tensorflow/core/util/env_var.h"
 #include "tensorflow/stream_executor/cuda/cuda_diagnostics.h"
 #include "tensorflow/stream_executor/gpu/asm_compiler.h"
 
@@ -150,16 +149,6 @@ Status NVPTXCompiler::OptimizeHloConvolutionCanonicalization(
   TF_RETURN_IF_ERROR(pipeline.Run(hlo_module).status());
 
   return Status::OK();
-}
-
-// TODO(cheshire): Duplication with gpu_conv_algorithm picker, figure out a
-// right way to share this.
-static bool RequireDeterminism() {
-  bool deterministic_ops = false;
-  TF_CHECK_OK(tensorflow::ReadBoolFromEnvVar("TF_DETERMINISTIC_OPS",
-                                             /*default_val=*/false,
-                                             &deterministic_ops));
-  return deterministic_ops;
 }
 
 Status NVPTXCompiler::OptimizeHloPostLayoutAssignment(
